Move most internal types to public types
